[API-93] Extract Home Assistant config/retry plumbing into config.ts.
