+++ api/data/home-assistant/index.ts
@@ -1,20 +1,6 @@
 import type { Zone } from '@/models';
 import { HttpResponseError, retry } from './retry';
-
-type SwitchService = 'turn_on' | 'turn_off';
-
-type HomeAssistantConfig = {
-    url: string;
-    token: string;
-};
-
-type RetryConfig = {
-    maxAttempts: number;
-    baseMs: number;
-};
-
-const DEFAULT_RETRY_MAX = 3;
-const DEFAULT_RETRY_BASE_MS = 1000;
+import { readConfig, readRetryConfig, type SwitchService } from './config';
 
 /**
  * Resolved switch state for a zone's relay. `'unknown'` covers the cases
@@ -313,33 +299,6 @@ async function sendRequest(endpoint: string, token: string, entityId: string, se
         console.error(message);
         throw new HttpResponseError(response.status, response.statusText, message);
     }
-}
-
-function readConfig(): HomeAssistantConfig {
-    const url = process.env.HA_URL;
-    const token = process.env.HA_TOKEN;
-    if (!url) throw new Error('home-assistant: HA_URL environment variable is required.');
-    if (!token) throw new Error('home-assistant: HA_TOKEN environment variable is required.');
-    return { url, token };
-}
-
-function readRetryConfig(): RetryConfig {
-    return {
-        maxAttempts: parsePositiveInt(process.env.HA_RETRY_MAX, DEFAULT_RETRY_MAX),
-        baseMs: parseNonNegativeInt(process.env.HA_RETRY_BASE_MS, DEFAULT_RETRY_BASE_MS),
-    };
-}
-
-function parsePositiveInt(raw: string | undefined, fallback: number): number {
-    if (raw === undefined) return fallback;
-    const parsed = Number.parseInt(raw, 10);
-    return Number.isFinite(parsed) && parsed >= 1 ? parsed : fallback;
-}
-
-function parseNonNegativeInt(raw: string | undefined, fallback: number): number {
-    if (raw === undefined) return fallback;
-    const parsed = Number.parseInt(raw, 10);
-    return Number.isFinite(parsed) && parsed >= 0 ? parsed : fallback;
 }
 
 function buildServiceUrl(baseUrl: string, service: SwitchService): string {
